fix:change place pages
