--- pages/index.tsx
@@ -1,15 +1,0 @@
-import type {NextPage} from 'next'
-import styles from '../src/styles/Home.module.css'
-import MainContainer from "../components/MainContainer";
-
-const Home: NextPage = () => {
-  return (
-      <MainContainer>
-    <div className={styles.container}>
-      <h1>HOME PAGE</h1>
-    </div>
-      </MainContainer>
-  )
-}
-
-export default Home
