+++ src/pages/myAccount.tsx
